v3.8.0

+++ Kotlin/Demo/app/src/main/java/io/chirp/sdkdemoapp/MainActivity.kt
@@ -136,7 +136,7 @@ class MainActivity : AppCompatActivity() {
             }
         }
 
-        chirpConnect.onSystemVolumeChanged { oldVolume: Int, newVolume: Int ->
+        chirpConnect.onSystemVolumeChanged { oldVolume: Float, newVolume: Float ->
             /**
              * onSystemVolumeChanged is called when the system volume is changed.
              *
